Break SearchPage and SearchForm into smaller components
