+++ talentsearch/resources/js/components/search/SearchForm.tsx
@@ -17,7 +17,6 @@ import {
   PoolCandidateFilter,
   PoolCandidateFilterInput,
 } from "../../api/generated";
-import EstimatedCandidates from "./EstimatedCandidates";
 
 const FilterBlock: React.FunctionComponent<{
   id: string;
@@ -34,7 +33,11 @@ const FilterBlock: React.FunctionComponent<{
       >
         {title}
       </h3>
-      <p data-h2-font-size="b(caption)" data-h2-margin="b(bottom, m)">
+      <p
+        data-h2-font-size="b(caption)"
+        data-h2-margin="b(bottom, m)"
+        data-h2-padding="b(right, xl)"
+      >
         {text}
       </p>
       {children && <div style={{ maxWidth: "30rem" }}>{children}</div>}
@@ -42,108 +45,12 @@ const FilterBlock: React.FunctionComponent<{
   );
 };
 
-const PoolBlock: React.FunctionComponent<{
-  pool: Pool;
-  totalEstimatedCandidates: number;
-  setPoolIdValue: (poolId: string) => void;
-  handleSubmit: () => void;
-}> = ({ pool, totalEstimatedCandidates, setPoolIdValue, handleSubmit }) => {
-  const intl = useIntl();
-  const locale = getLocale(intl);
-
-  function span(msg: string) {
-    return (
-      <span data-h2-font-weight="b(700)" data-h2-font-color="b(lightpurple)">
-        {msg}
-      </span>
-    );
-  }
-
-  return (
-    <div
-      data-h2-shadow="b(m)"
-      data-h2-border="b(lightnavy, left, solid, l)"
-      data-h2-margin="b(top, s) b(bottom, m)"
-      data-h2-flex-grid="b(middle, contained, flush, xl)"
-    >
-      <div
-        data-h2-flex-item="b(1of1) m(1of2)"
-        style={{ padding: "0", paddingLeft: "1rem" }}
-      >
-        <p data-h2-margin="b(bottom, none)" data-h2-font-weight="b(700)">
-          {pool.name?.[locale]
-            ? pool.name?.[locale]
-            : intl.formatMessage({
-                defaultMessage: "N/A",
-                description: "Text shown when the filter was not selected",
-              })}
-        </p>
-        <p
-          data-h2-margin="b(top, xxs) b(bottom, m)"
-          data-h2-font-weight="b(100)"
-        >
-          {intl.formatMessage(
-            {
-              defaultMessage:
-                "There are <span>{totalEstimatedCandidates}</span> matching candidates in this pool",
-              description:
-                "Message for total estimated candidates box next to search form.",
-            },
-            {
-              span,
-              totalEstimatedCandidates,
-            },
-          )}
-        </p>
-        <p data-h2-margin="b(bottom, none)" data-h2-font-size="b(caption)">
-          {intl.formatMessage({ defaultMessage: "Pool Owner" })}:{" "}
-          {pool.owner?.firstName
-            ? pool.owner?.firstName
-            : intl.formatMessage({
-                defaultMessage: "N/A",
-                description: "Text shown when the filter was not selected",
-              })}{" "}
-          {pool.owner?.lastName
-            ? pool.owner?.lastName
-            : intl.formatMessage({
-                defaultMessage: "N/A",
-                description: "Text shown when the filter was not selected",
-              })}
-        </p>
-        <p data-h2-margin="b(bottom, s)" data-h2-font-size="b(caption)">
-          {pool.description?.[locale]
-            ? pool.description?.[locale]
-            : intl.formatMessage({
-                defaultMessage: "N/A",
-                description: "Text shown when the filter was not selected",
-              })}
-        </p>
-      </div>
-      <div
-        data-h2-flex-item="b(1of1) m(1of2)"
-        data-h2-display="b(flex)"
-        data-h2-justify-content="b(center) m(flex-end)"
-      >
-        <Button
-          color="cta"
-          mode="solid"
-          onClick={() => {
-            // Set the poolId state
-            setPoolIdValue(pool.id);
-            // Fire submit handler with form data and save to history state api.
-            handleSubmit();
-          }}
-        >
-          {intl.formatMessage({
-            defaultMessage: "Request Candidates",
-            description:
-              "Button link message on search page that takes user to the request form.",
-          })}
-        </Button>
-      </div>
-    </div>
-  );
-};
+function mapIdToValue<T extends { id: Id }>(objs: T[]): Map<Id, T> {
+  return objs.reduce((map, obj) => {
+    map.set(obj.id, obj);
+    return map;
+  }, new Map());
+}
 
 type Option<V> = { value: V; label: string };
 type FormValues = Pick<
@@ -155,20 +62,13 @@ type FormValues = Pick<
   operationalRequirements: Id[] | undefined;
   employmentEquity: string[] | undefined;
   educationRequirement: "has_diploma" | "no_diploma";
+  poolId: string;
 };
-
 export interface SearchFormProps {
   classifications: Classification[];
   cmoAssets: CmoAsset[];
   operationalRequirements: OperationalRequirement[];
   updateCandidateFilter: (filter: PoolCandidateFilterInput) => void;
-}
-
-function mapIdToValue<T extends { id: Id }>(objs: T[]): Map<Id, T> {
-  return objs.reduce((map, obj) => {
-    map.set(obj.id, obj);
-    return map;
-  }, new Map());
 }
 
 export const SearchForm: React.FunctionComponent<SearchFormProps> = ({
@@ -179,7 +79,6 @@ export const SearchForm: React.FunctionComponent<SearchFormProps> = ({
 }) => {
   const intl = useIntl();
   const locale = getLocale(intl);
-  const location = useLocation();
 
   const classificationMap = useMemo(
     () => mapIdToValue(classifications),
@@ -210,13 +109,13 @@ export const SearchForm: React.FunctionComponent<SearchFormProps> = ({
       values.languageAbility === "BILINGUAL"
         ? { languageAbility: values.languageAbility }
         : {}), // Ensure null in FormValues is converted to undefined
-      workRegions: values.workRegions,
+      workRegions: values.workRegions || [],
     }),
     [classificationMap, assetMap, requirementMap],
   );
 
   const methods = useForm<FormValues>();
-  const { watch } = methods;
+  const { watch, getValues } = methods;
 
   // Whenever form values change (with some debounce allowance), call updateCandidateFilter
   const formValues = watch();
@@ -264,72 +163,9 @@ export const SearchForm: React.FunctionComponent<SearchFormProps> = ({
     [operationalRequirements, locale],
   );
 
-  const onSubmit: SubmitHandler<FormValues> = async (data: FormValues) => {
-    // Build the poolCandidateFilter that will be stored into the history.state api.
-    const poolCandidateFilter: PoolCandidateFilter = {
-      id: "",
-      classifications:
-        data.classifications && data.classifications.length > 0
-          ? data.classifications?.map((classification) => {
-              const newClassification = classifications.find(
-                ({ id }) => id === classification,
-              );
-              return newClassification;
-            })
-          : [],
-      cmoAssets:
-        data.cmoAssets && data.cmoAssets.length > 0
-          ? data.cmoAssets.map((cmoAsset) => {
-              const newCmoAsset = cmoAssets.find(({ id }) => id === cmoAsset);
-              return newCmoAsset;
-            })
-          : [],
-      operationalRequirements:
-        data.operationalRequirements && data.operationalRequirements.length > 0
-          ? data.operationalRequirements.map((operationalRequirement) => {
-              const newOperationalRequirement = operationalRequirements.find(
-                ({ id }) => id === operationalRequirement,
-              );
-              return newOperationalRequirement;
-            })
-          : [],
-      pools: pools.filter(({ id }) => id === data.poolId),
-      hasDiploma: !!Number(data.hasDiploma),
-      hasDisability: data.employmentEquity
-        ? data.employmentEquity?.includes("hasDisability")
-        : false,
-      isIndigenous: data.employmentEquity
-        ? data.employmentEquity?.includes("isIndigenous")
-        : false,
-      isVisibleMinority: data.employmentEquity
-        ? data.employmentEquity?.includes("isVisibleMinority")
-        : false,
-      isWoman: data.employmentEquity
-        ? data.employmentEquity?.includes("isWoman")
-        : false,
-      languageAbility:
-        data.languageAbility === "null" ? null : data.languageAbility,
-      workRegions: data.workRegions,
-    };
-    return pushToStateThenNavigate(requestPath(), {
-      poolCandidateFilter,
-      initialValues: data,
-    });
-  };
-
-  const useHandleSubmit = () => handleSubmit(onSubmit)();
-
-  function span(msg: string): JSX.Element {
-    return <span data-h2-font-color="b(lightpurple)">{msg}</span>;
-  }
-
-  function a(msg: string): JSX.Element {
-    return (
-      <a href="/search" data-h2-font-weight="b(700)">
-        {msg}
-      </a>
-    );
-  }
+  console.log(
+    getValues("classifications")?.map((id) => classificationMap.get(id)),
+  );
 
   return (
     <FormProvider {...methods}>
@@ -572,50 +408,50 @@ export const SearchForm: React.FunctionComponent<SearchFormProps> = ({
   );
 };
 
-export const SearchFormApi: React.FunctionComponent = () => {
-  const intl = useIntl();
-  const [lookupResult] = useGetSearchFormDataQuery();
-  const { data: lookupData, fetching, error } = lookupResult;
-  const classifications: Classification[] | [] =
-    lookupData?.classifications.filter(notEmpty) ?? [];
-  const cmoAssets: CmoAsset[] = lookupData?.cmoAssets.filter(notEmpty) ?? [];
-  const operationalRequirements: OperationalRequirement[] =
-    lookupData?.operationalRequirements.filter(notEmpty) ?? [];
-  const pools: Pool[] = lookupData?.pools.filter(notEmpty) ?? [];
+// export const SearchFormApi: React.FunctionComponent = () => {
+//   const intl = useIntl();
+//   const [lookupResult] = useGetSearchFormDataQuery();
+//   const { data: lookupData, fetching, error } = lookupResult;
+//   const classifications: Classification[] | [] =
+//     lookupData?.classifications.filter(notEmpty) ?? [];
+//   const cmoAssets: CmoAsset[] = lookupData?.cmoAssets.filter(notEmpty) ?? [];
+//   const operationalRequirements: OperationalRequirement[] =
+//     lookupData?.operationalRequirements.filter(notEmpty) ?? [];
+//   const pools: Pool[] = lookupData?.pools.filter(notEmpty) ?? [];
 
-  if (fetching)
-    return (
-      <div
-        data-h2-display="b(flex)"
-        data-h2-justify-content="b(center)"
-        data-h2-align-items="b(center)"
-        style={{ minHeight: "20rem" }}
-      >
-        <p>{intl.formatMessage(commonMessages.loadingTitle)}</p>
-      </div>
-    );
-  if (error)
-    return (
-      <div
-        data-h2-display="b(flex)"
-        data-h2-justify-content="b(center)"
-        data-h2-align-items="b(center)"
-        style={{ minHeight: "20rem" }}
-      >
-        <p>
-          {intl.formatMessage(commonMessages.loadingError)}
-          {error.message}
-        </p>
-      </div>
-    );
+//   if (fetching)
+//     return (
+//       <div
+//         data-h2-display="b(flex)"
+//         data-h2-justify-content="b(center)"
+//         data-h2-align-items="b(center)"
+//         style={{ minHeight: "20rem" }}
+//       >
+//         <p>{intl.formatMessage(commonMessages.loadingTitle)}</p>
+//       </div>
+//     );
+//   if (error)
+//     return (
+//       <div
+//         data-h2-display="b(flex)"
+//         data-h2-justify-content="b(center)"
+//         data-h2-align-items="b(center)"
+//         style={{ minHeight: "20rem" }}
+//       >
+//         <p>
+//           {intl.formatMessage(commonMessages.loadingError)}
+//           {error.message}
+//         </p>
+//       </div>
+//     );
 
-  return (
-    <SearchForm
-      classifications={classifications}
-      cmoAssets={cmoAssets}
-      operationalRequirements={operationalRequirements}
-      pools={pools}
-      totalEstimatedCandidates={0} // TODO: Replace with number of candidates queried from api.
-    />
-  );
-};
+//   return (
+//     <SearchForm
+//       classifications={classifications}
+//       cmoAssets={cmoAssets}
+//       operationalRequirements={operationalRequirements}
+//       pools={pools}
+//       totalEstimatedCandidates={0} // TODO: Replace with number of candidates queried from api.
+//     />
+//   );
+// };
